+++ lib/src/ktlint/java/com/diffplug/spotless/glue/ktlint/KtlintFormatterFunc.java
@@ -17,18 +17,26 @@ package com.diffplug.spotless.glue.ktlint;
 
 import java.io.File;
 import java.util.ArrayList;
+import java.util.Arrays;
 import java.util.List;
 import java.util.Map;
+import java.util.Objects;
+import java.util.stream.Collectors;
+import java.util.stream.Stream;
 
 import com.pinterest.ktlint.core.KtLint;
-import com.pinterest.ktlint.core.KtLint.Params;
+import com.pinterest.ktlint.core.KtLint.ExperimentalParams;
 import com.pinterest.ktlint.core.LintError;
 import com.pinterest.ktlint.core.RuleSet;
+import com.pinterest.ktlint.core.api.DefaultEditorConfigProperties;
+import com.pinterest.ktlint.core.api.EditorConfigOverride;
+import com.pinterest.ktlint.core.api.UsesEditorConfigProperties;
 import com.pinterest.ktlint.ruleset.experimental.ExperimentalRuleSetProvider;
 import com.pinterest.ktlint.ruleset.standard.StandardRuleSetProvider;
 
 import com.diffplug.spotless.FormatterFunc;
 
+import kotlin.Pair;
 import kotlin.Unit;
 import kotlin.jvm.functions.Function2;
 
@@ -38,8 +46,13 @@ public class KtlintFormatterFunc implements FormatterFunc.NeedsFile {
 	private final Map<String, String> userData;
 	private final Function2<? super LintError, ? super Boolean, Unit> formatterCallback;
 	private final boolean isScript;
+	private final EditorConfigOverride editorConfigOverride;
 
-	public KtlintFormatterFunc(boolean isScript, boolean useExperimental, Map<String, String> userData) {
+	/**
+	 * Non-empty editorConfigOverrideMap requires KtLint 0.45.2.
+	 */
+	public KtlintFormatterFunc(boolean isScript, boolean useExperimental, Map<String, String> userData,
+			Map<String, Object> editorConfigOverrideMap) {
 		rulesets = new ArrayList<>();
 		rulesets.add(new StandardRuleSetProvider().get());
 
@@ -49,6 +62,46 @@ public class KtlintFormatterFunc implements FormatterFunc.NeedsFile {
 		this.userData = userData;
 		formatterCallback = new FormatterCallback();
 		this.isScript = isScript;
+
+		if (editorConfigOverrideMap.isEmpty()) {
+			this.editorConfigOverride = null;
+		} else {
+			this.editorConfigOverride = createEditorConfigOverride(editorConfigOverrideMap);
+		}
+	}
+
+	/**
+	 * Create EditorConfigOverride from user provided parameters.
+	 * Calling this method requires KtLint 0.45.2.
+	 */
+	private EditorConfigOverride createEditorConfigOverride(Map<String, Object> editorConfigOverrideMap) {
+		// Get properties from rules in the rule sets
+		Stream<UsesEditorConfigProperties.EditorConfigProperty<?>> ruleProperties = rulesets.stream()
+				.flatMap(ruleSet -> Arrays.stream(ruleSet.getRules()))
+				.filter(rule -> rule instanceof UsesEditorConfigProperties)
+				.flatMap(rule -> ((UsesEditorConfigProperties) rule).getEditorConfigProperties().stream());
+
+		// Create a mapping of properties to their names based on rule properties and default properties
+		Map<String, UsesEditorConfigProperties.EditorConfigProperty<?>> supportedProperties = Stream
+				.concat(ruleProperties, DefaultEditorConfigProperties.INSTANCE.getDefaultEditorConfigProperties().stream())
+				.distinct()
+				.collect(Collectors.toMap(property -> property.getType().getName(), property -> property));
+
+		// Create config properties based on provided property names and values
+		@SuppressWarnings("unchecked")
+		Pair<UsesEditorConfigProperties.EditorConfigProperty<?>, ?>[] properties = editorConfigOverrideMap.entrySet().stream()
+				.map(entry -> {
+					UsesEditorConfigProperties.EditorConfigProperty<?> property = supportedProperties.get(entry.getKey());
+					if (property != null) {
+						return new Pair<>(property, entry.getValue());
+					} else {
+						return null;
+					}
+				})
+				.filter(Objects::nonNull)
+				.toArray(Pair[]::new);
+
+		return EditorConfigOverride.Companion.from(properties);
 	}
 
 	static class FormatterCallback implements Function2<LintError, Boolean, Unit> {
@@ -63,14 +116,31 @@ public class KtlintFormatterFunc implements FormatterFunc.NeedsFile {
 
 	@Override
 	public String applyWithFile(String unix, File file) throws Exception {
-		return KtLint.INSTANCE.format(new Params(
-				file.getName(),
-				unix,
-				rulesets,
-				userData,
-				formatterCallback,
-				isScript,
-				null,
-				false));
+
+		if (editorConfigOverride != null) {
+			// Use ExperimentalParams with EditorConfigOverride which requires KtLint 0.45.2
+			return KtLint.INSTANCE.format(new ExperimentalParams(
+					file.getName(),
+					unix,
+					rulesets,
+					userData,
+					formatterCallback,
+					isScript,
+					null,
+					false,
+					editorConfigOverride,
+					false));
+		} else {
+			// Use Params for backward compatibility
+			return KtLint.INSTANCE.format(new KtLint.Params(
+					file.getName(),
+					unix,
+					rulesets,
+					userData,
+					formatterCallback,
+					isScript,
+					null,
+					false));
+		}
 	}
 }